+++ mlrun/api/crud/model_monitoring/deployment.py
@@ -43,7 +43,7 @@ _MONITORING_ORIGINAL_BATCH_FUNCTION_PATH = (
 )
 
 _MONITORING_APPLICATION_BATCH_FUNCTION_PATH = (
-    _MODEL_MONITORING_COMMON_PATH / "model_monitoring_application.py"
+    _MODEL_MONITORING_COMMON_PATH / "model_monitoring_batch_application.py"
 )
 
 _MONITORING_WRITER_FUNCTION_PATH = (
